requirements.txt
Django==1.11.4
djangorestframework==3.6.4
Pillow==4.2.1
psycopg2==2.7.3.1
